combined otp form
