+++ apps/login/ui/LoginPasskey.tsx
@@ -6,6 +6,7 @@ import { coerceToArrayBuffer, coerceToBase64Url } from "#/utils/base64";
 import { Button, ButtonVariants } from "./Button";
 import Alert from "./Alert";
 import { Spinner } from "./Spinner";
+import { Checks } from "@zitadel/server";
 
 // either loginName or sessionId must be provided
 type Props = {
@@ -100,7 +101,9 @@ export default function LoginPasskey({
         loginName,
         sessionId,
         organization,
-        webAuthN: { credentialAssertionData: data },
+        checks: {
+          webAuthN: { credentialAssertionData: data },
+        } as Checks,
         authRequestId,
       }),
     });
